rsc_table: add notifyid fields to the virtio resource entries
Linux needs a way to notify SYS/BIOS if a vring has pending messages, or if the status of a certain vdev changes.

This is done by sending the notifyid of that specific vring/vdev over the mailbox payload.

This patch adds a notifyid member to the relevant resource entries, which will be used to synchronize the notify index of that specific object.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>