Resources: Adjust the physical address of the vrings
The memory allocation on the host-side has switched to use the
Contiguous Memory Allocator (CMA). All the memory regions needing
DDR memory are allocated by the CMA, and the IPC sections are the
first on the per-device CMA region. However, the IPC sections also
needs to be mapped into the MMU, and hence are specified as DEVMEM
entries, and need a physical address matching the start of the CMA
regions. The current alignment requirements for CMA is 8MB, and
this patch corrects the addresses as per the new host-side CMA
reservations.

Signed-off-by: Subramaniam C.A <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>