Resources: Adapt to a new TLV-style resource table
The previous resource table is defined as an array of different
resources having the same structure. This approach was very
limiting - all members had to have the same size and members,
and this didn't work quite well since some members are very
small, where others quite big.

The resource table hence is moved to a new type-value based
approach, where each resource entry has its own dedicated
structure. The resource table begins with a simple header
which contains a version field and the number of entries and
their offsets within the table. The version field helps the
host a means to verify the image format they are expecting,
and the offsets help the parsing code to do fast browsing
through the table.

The IpcMemory module has also fixed up to work with the new
resource table format, and continue to provide the memory
translation functions for converting between device/virtual
memory and the actual physical memory.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>