Resources: Fix insufficient memory allocation for IpcMemory_RscTable
The IpcMemory_RscTable structure definition is using an open-ended
array declaration for the 'offset' element for which the TI linker
has failed to assign any memory. This is causing an erroneous offset
translation in the IpcMemory_getEntry function. The invalid offset
can result in a random value and thereby cause an exception in the
IpcMemory_virtToPhys and IpcMemory_physToVirt API. The 'offset'
variable is now defined as an array with 1 element to make sure the
structure object is allocated enough memory and allow the 'offset'
to be dereferenced correctly.

Signed-off-by: Subramaniam C.A <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>