Build:  Internal rework of Makefile and build scripts
Removed the ti.build package and updated _some_ package.bld scripts to no
longer leverage it.  Also explicitly list the packages to build
(in top-level Makefile) as not all package.bld scripts were updated.  Those
that were not updated are not needed for OMAP-L138.

There is minor cleanup of a few sources files - inspired by the extra remarks
and warnings being enabled in the toolchain.