TCI6614: Allowed Interrupt to be shared with TransportSHM and other clients.
On keystone, though there is only one IPC interrupt line, there is a register
which can indication up to 27 different source ID bits.

This allows muxing interrupts through the register, over this one interrupt
line.

The Interrupt module for tci6614 has been updated to allow such muxing,
which will also enable it to share the interrupt line with other transports.

The other transports will need to be configured to pass in the intInfo
localIntId value corresponding to the remote proc source ID.

Signed-off-by: G Anthony <a0783926@ti.com>